requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crypto==1.4.1
cryptography==40.0.2
Faker==18.9.0
idna==3.4
Naked==0.1.32
numpy==1.24.3
pandas==2.0.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
shellescape==3.8.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
